requirements/testing.txt
coverage==4.0.3
django-nose==1.4.5
nose==1.3.7
nose-exclude==0.4.1
pycodestyle==2.3.1
pylint==1.8.4
mock==2.0.0
Faker==0.8.6
python-coveralls==2.9.1
coveralls==1.2.0
freezegun==0.3.11
